refator: lista de logs movida para outro arquivo e funcionalidades separadas
