+++ src/feature/Alteracoes/interfaces/index.ts
@@ -1,0 +1,7 @@
+export interface IHistorico {
+    id: number
+    nome: string
+    mensagem: string
+    acao: string
+    data_criacao: string
+}
